requirements.txt
# GAA Framework - Requirements
# Generación Automática de Algoritmos

# ============================================================================
# CORE DEPENDENCIES
# ============================================================================
numpy>=1.24.0
pyyaml>=6.0

# ============================================================================
# METAHEURISTICS & OPTIMIZATION
# ============================================================================
# Para implementar algoritmos evolutivos (GP, GA)
deap>=1.4.0
# Para Simulated Annealing y otras metaheurísticas
scipy>=1.10.0

# ============================================================================
# AST & CODE GENERATION
# ============================================================================
# Para manipulación de AST Python
astor>=0.8.1
# Para análisis de código
astroid>=2.15.0

# ============================================================================
# DATA PROCESSING
# ============================================================================
pandas>=2.0.0
# Para grafos (Graph Coloring Problem)
networkx>=3.0

# ============================================================================
# VISUALIZATION
# ============================================================================
matplotlib>=3.7.0
seaborn>=0.12.0
# Para visualización de AST
graphviz>=0.20.0

# ============================================================================
# TESTING & VALIDATION
# ============================================================================
pytest>=7.4.0
pytest-cov>=4.1.0

# ============================================================================
# DEVELOPMENT TOOLS
# ============================================================================
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# ============================================================================
# UTILITIES
# ============================================================================
tqdm>=4.65.0  # Progress bars
colorama>=0.4.6  # Colored terminal output
